cambios a watchdog de la maquina de estado

+++ SleepingEndDevice/Source/app_sleeping_enddevice.h
@@ -83,7 +83,7 @@ typedef enum
 
 typedef struct
 {
-	uint16						systemStrikes;
+	uint8						systemStrikes;
 	bool						isConfigured;
 	uint16						samplePeriod;
 	uint16						channelAValue;
